libmsg crate for message interfaces to compositor and engine
