+++ src/components/gfx/gfx.rc
@@ -15,6 +15,7 @@ extern mod stb_image;
 extern mod std;
 extern mod servo_net (name = "net");
 extern mod servo_util (name = "util");
+extern mod servo_msg (name = "msg");
 
 // Eventually we would like the shaper to be pluggable, as many operating systems have their own
 // shapers. For now, however, this is a hard dependency.
@@ -39,7 +40,6 @@ priv mod render_context;
 
 // Rendering
 pub mod color;
-pub mod compositor;
 pub mod display_list;
 pub mod geometry;
 pub mod render_task;
