+++ src/components/gfx/render_context.rs
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-use compositor::LayerBuffer;
+use servo_msg::compositor::LayerBuffer;
 use font_context::FontContext;
 use geometry::Au;
 use opts::Opts;
