+++ src/components/main/engine.rs
@@ -10,8 +10,8 @@ use core::comm::Port;
 use gfx::opts::Opts;
 use gfx::render_task::RenderChan;
 use gfx::render_task;
-use script::compositor_interface::{ScriptListener, ReadyState};
-use script::engine_interface::{EngineChan, ExitMsg, LoadUrlMsg, Msg};
+use servo_msg::compositor::{ScriptListener, ReadyState};
+use servo_msg::engine::{EngineChan, ExitMsg, LoadUrlMsg, Msg};
 use script::layout_interface::LayoutChan;
 use script::layout_interface;
 use script::script_task::{ExecuteMsg, LoadMsg, ScriptMsg, ScriptContext, ScriptChan};
