+++ src/components/main/platform/common/glut_windowing.rs
@@ -15,11 +15,11 @@ use alert::{Alert, AlertMethods};
 use core::libc::c_int;
 use geom::point::Point2D;
 use geom::size::Size2D;
-use gfx::compositor::{IdleRenderState, RenderState, RenderingRenderState};
+use servo_msg::compositor::{IdleRenderState, RenderState, RenderingRenderState};
+use servo_msg::compositor::{FinishedLoading, Loading, PerformingLayout, ReadyState};
 use glut::glut::{ACTIVE_CTRL, DOUBLE, HAVE_PRECISE_MOUSE_WHEEL, WindowHeight, WindowWidth};
 use glut::glut;
 use glut::machack;
-use script::compositor_interface::{FinishedLoading, Loading, PerformingLayout, ReadyState};
 
 static THROBBER: [char, ..8] = [ '⣾', '⣽', '⣻', '⢿', '⡿', '⣟', '⣯', '⣷' ];
