--- src/components/script/compositor_interface.rs
@@ -1,21 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-//! The high-level interface from script to compositor. Using this abstract interface helps reduce
-/// coupling between these two components
-
-pub enum ReadyState {
-    /// Informs the compositor that a page is loading. Used for setting status
-    Loading,
-    /// Informs the compositor that a page is performing layout. Used for setting status
-    PerformingLayout,
-    /// Informs the compositor that a page is finished loading. Used for setting status
-    FinishedLoading,
-}
-
-/// The interface used by the script task to tell the compositor to update its ready state,
-/// which is used in displaying the appropriate message in the window's title.
-pub trait ScriptListener : Clone {
-    fn set_ready_state(&self, ReadyState);
-}
